+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 ## Unreleased
+### 6.1.2
+
+#### Notable Changes
+
+- api/news:
+   - fix out of order related documents
+   - fix out of order designers/engineers/photographers/writers
 
 ## Released
 ### 6.1.1 (Current), 2020-09-03